dump: Adjust sampling time based on elapsed time for flame graph
When user doesn't give a sample time, it used number of calls as
sample time for flame graph.  But it may lead to a confusing result
that graph can show small and frequent functions more than actual
big functions that consume most of the time.

So it'd be better to use specify a sample time.  But it might be hard
for user to find the right granularity.  Since we have elapsed time of
the recorded data, we can calculate the sample time based on the
target sampling frequency.  Right now, I randomly chose 1 million as
the target (max) sample count.

Closed: #1257

Signed-off-by: Namhyung Kim <namhyung@gmail.com>